Add initial framework for the LCM type
